// web/src/form/questions/response/__generated__/ResponseModeCheckboxQuestionResponseQuery.graphql.ts
export type ResponseModeCheckboxQuestionResponseQuery$variables = {
  formInstanceID: string;
  questionID: string;
};
export type ResponseModeCheckboxQuestionResponseQuery$data = {
  readonly questionResponses: {
    readonly edges: ReadonlyArray<{
      readonly node: {
        readonly __typename: "QuestionResponse";
        readonly id: string;
        readonly value: string;
      } | null | undefined;
    } | null | undefined> | null | undefined;
  };
};
export type ResponseModeCheckboxQuestionResponseQuery = {
  response: ResponseModeCheckboxQuestionResponseQuery$data;
  variables: ResponseModeCheckboxQuestionResponseQuery$variables;
};